requirements.txt
# ── Interactive Brokers ────────────────────────────────────────────────────
ib_insync==0.9.86

# ── Data / ML ─────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
alpaca-trade-api==3.3.0

# ── Web / API ──────────────────────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
aiohttp==3.9.5
python-dotenv==1.0.1
requests==2.32.3

# ── Telegram ──────────────────────────────────────────────────────────────
python-telegram-bot==21.3

# ── C++ bindings ──────────────────────────────────────────────────────────
pybind11==2.12.0

# ── Utilities ─────────────────────────────────────────────────────────────
pytz==2024.1
tabulate==0.9.0
